feat: streamline wrapper into screen
